# Fix splat args