--- a/yjit/src/stats.rs
+++ b/yjit/src/stats.rs
@@ -193,9 +193,9 @@ make_counters! {
     send_getter_arity,
     send_se_cf_overflow,
     send_se_protected_check_failed,
-    send_splatarray_rhs_too_small,
+    send_splatarray_length_not_equal,
     send_splat_not_array,
-    send_args_splat_variadic,
+    send_args_splat_non_iseq,
     send_iseq_ruby2_keywords,
 
     traced_cfunc_return,