@maximecb 2022-09-15
Looking good so far 👍  Once all the tests are passing, can you post the stats from railsbench?