@k0kubun 2022-09-16
master CI is failing fairly frequently, so I'd like to merge this.

> Once all the tests are passing, can you post the stats from railsbench?

## before

<details>

```
$ ./run_benchmarks.rb railsbench -e "$(rbenv ruby yjit-$arch) --yjit-stats"
Running benchmark "railsbench" (1/1)
/Users/k0kubun/.rbenv/versions/yjit-arm64/bin/ruby --yjit-stats -I ./harness benchmarks/railsbench/benchmark.rb
ruby 3.2.0dev (2022-09-15T14:14:27Z master cc7f6fe734) +YJIT [arm64-darwin21]
Command: bundle install --quiet
Command: bin/rails db:migrate db:seed
Calling `DidYouMean::SPELL_CHECKERS.merge!(error_name => spell_checker)' has been deprecated. Please call `DidYouMean.correct_error(error_name, spell_checker)' instead.
Deleted all 100 posts
Creating 100 posts....................................................................................................
itr #1: 2429ms
itr #2: 1265ms
itr #3: 1280ms
itr #4: 1279ms
itr #5: 1279ms
itr #6: 1225ms
itr #7: 1235ms
itr #8: 1244ms
itr #9: 1266ms
itr #10: 1391ms
itr #11: 1330ms
itr #12: 1272ms
itr #13: 1354ms
itr #14: 1304ms
itr #15: 1358ms
itr #16: 1264ms
itr #17: 1332ms
itr #18: 1332ms
itr #19: 1301ms
itr #20: 1327ms
itr #21: 1281ms
itr #22: 1289ms
itr #23: 1252ms
itr #24: 1392ms
itr #25: 1315ms
Average of last 10, non-warmup iters: 1308ms
***YJIT: Printing YJIT statistics on exit***
method call exit reasons:
                block_arg    6418043 (33.0%)
    optimized_method_call    4262130 (21.9%)
      iseq_complex_callee    2931224 (15.1%)
                  bmethod    2348191 (12.1%)
              iseq_zsuper    1151795 ( 5.9%)
      iseq_ruby2_keywords     885835 ( 4.6%)
    optimized_method_send     718524 ( 3.7%)
      args_splat_variadic     398574 ( 2.1%)
                 kw_splat     314184 ( 1.6%)
           refined_method        854 ( 0.0%)
    cfunc_ruby_array_varg        487 ( 0.0%)
         iseq_arity_error        101 ( 0.0%)
                 keywords         92 ( 0.0%)
            zsuper_method          7 ( 0.0%)
invokesuper exit reasons:
         block     115235 (69.7%)
    me_changed      50056 (30.3%)
leave exit reasons:
        interp_return   41012078 (96.3%)
    start_pc_non_zero    1578476 ( 3.7%)
         se_interrupt        220 ( 0.0%)
getblockparamproxy exit reasons:
    block_param_modified         97 (100.0%)
getinstancevariable exit reasons:
    idx_out_of_range          2 (100.0%)
setinstancevariable exit reasons:
    (all relevant counters are zero)
opt_aref exit reasons:
    (all relevant counters are zero)
expandarray exit reasons:
            splat     249235 (100.0%)
    rhs_too_small         14 ( 0.0%)
opt_getinlinecache exit reasons:
    miss          5 (100.0%)
invalidation reasons:
          method_lookup        227 (44.2%)
    constant_state_bump        199 (38.7%)
       constant_ic_fill         88 (17.1%)
bindings_allocations:         162
bindings_set:                   0
compiled_iseq_count:         5628
compiled_block_count:       23200
invalidation_count:           514
constant_state_bumps:           0
inline_code_size:         3983152
outlined_code_size:       3202364
num_gc_obj_refs:            16269
total_exit_count:        63906767
total_insns_count:     2044996850
vm_insns_count:         280919091
yjit_insns_count:      1786972448
ratio_in_yjit:              86.3%
avg_len_in_yjit:             27.6
Top-20 most frequent exit ops (100.0% of exits):
    opt_send_without_block:   10451262 (45.6%)
                      send:    6600840 (28.8%)
               invokesuper:    2763913 (12.1%)
               invokeblock:    1603120 (7.0%)
               getconstant:     432849 (1.9%)
                  opt_aref:     350098 (1.5%)
                     throw:     259112 (1.1%)
               expandarray:     249249 (1.1%)
             setlocal_WC_0:     149541 (0.7%)
      opt_getconstant_path:      32602 (0.1%)
               opt_empty_p:        749 (0.0%)
       getinstancevariable:        415 (0.0%)
        getblockparamproxy:        245 (0.0%)
                     leave:        220 (0.0%)
                checkmatch:        193 (0.0%)
               objtostring:        139 (0.0%)
                 opt_nil_p:         30 (0.0%)
          opt_newarray_max:         29 (0.0%)
                      once:         27 (0.0%)
             setblockparam:         15 (0.0%)
Total time spent benchmarking: 38s

end_time: 2022-09-16 13:30:09 JST (+0900)
/Users/k0kubun/.rbenv/versions/yjit-arm64/bin/ruby --yjit-stats: ruby 3.2.0dev (2022-09-15T14:14:27Z master cc7f6fe734) +YJIT [arm64-darwin21]

----------  --------------------------------------------------------------------  ----------
bench       /Users/k0kubun/.rbenv/versions/yjit-arm64/bin/ruby --yjit-stats (ms)  stddev (%)
railsbench  1308.9                                                                2.9
----------  --------------------------------------------------------------------  ----------
```

</details>

## after

<details>

```
$ ./run_benchmarks.rb railsbench -e "$(rbenv ruby yjit-$arch) --yjit-stats"
Running benchmark "railsbench" (1/1)
/Users/k0kubun/.rbenv/versions/yjit-arm64/bin/ruby --yjit-stats -I ./harness benchmarks/railsbench/benchmark.rb
ruby 3.2.0dev (2022-09-15T20:19:50Z splat-args-fix e2a641724a) +YJIT [arm64-darwin21]
Command: bundle install --quiet
Command: bin/rails db:migrate db:seed
Calling `DidYouMean::SPELL_CHECKERS.merge!(error_name => spell_checker)' has been deprecated. Please call `DidYouMean.correct_error(error_name, spell_checker)' instead.
Deleted all 100 posts
Creating 100 posts....................................................................................................
itr #1: 2523ms
itr #2: 1217ms
itr #3: 1260ms
itr #4: 1251ms
itr #5: 1253ms
itr #6: 1242ms
itr #7: 1261ms
itr #8: 1253ms
itr #9: 1243ms
itr #10: 1243ms
itr #11: 1213ms
itr #12: 1212ms
itr #13: 1242ms
itr #14: 1256ms
itr #15: 1243ms
itr #16: 1238ms
itr #17: 1213ms
itr #18: 1257ms
itr #19: 1245ms
itr #20: 1234ms
itr #21: 1217ms
itr #22: 1213ms
itr #23: 1243ms
itr #24: 1252ms
itr #25: 1250ms
Average of last 10, non-warmup iters: 1236ms
***YJIT: Printing YJIT statistics on exit***
method call exit reasons:
                block_arg    6418050 (33.0%)
    optimized_method_call    4262129 (21.9%)
      iseq_complex_callee    2931229 (15.1%)
                  bmethod    2348191 (12.1%)
              iseq_zsuper    1151797 ( 5.9%)
      iseq_ruby2_keywords     885835 ( 4.6%)
    optimized_method_send     718526 ( 3.7%)
         args_splat_cfunc     348145 ( 1.8%)
                 kw_splat     314184 ( 1.6%)
      args_splat_non_iseq      50435 ( 0.3%)
           refined_method        854 ( 0.0%)
    cfunc_ruby_array_varg        487 ( 0.0%)
                 keywords         92 ( 0.0%)
         iseq_arity_error          7 ( 0.0%)
            zsuper_method          7 ( 0.0%)
invokesuper exit reasons:
         block     115235 (69.7%)
    me_changed      50056 (30.3%)
leave exit reasons:
        interp_return   41007758 (96.3%)
    start_pc_non_zero    1578476 ( 3.7%)
         se_interrupt        210 ( 0.0%)
getblockparamproxy exit reasons:
    block_param_modified         97 (100.0%)
getinstancevariable exit reasons:
    idx_out_of_range          2 (100.0%)
setinstancevariable exit reasons:
    (all relevant counters are zero)
opt_aref exit reasons:
    (all relevant counters are zero)
expandarray exit reasons:
            splat     249240 (100.0%)
    rhs_too_small         14 ( 0.0%)
opt_getinlinecache exit reasons:
    miss          5 (100.0%)
invalidation reasons:
          method_lookup        227 (44.1%)
    constant_state_bump        200 (38.8%)
       constant_ic_fill         88 (17.1%)
bindings_allocations:         162
bindings_set:                   0
compiled_iseq_count:         5628
compiled_block_count:       23176
invalidation_count:           515
constant_state_bumps:           0
inline_code_size:         3976536
outlined_code_size:       3198580
num_gc_obj_refs:            16262
total_exit_count:        63902245
total_insns_count:     2041478759
vm_insns_count:         283788291
yjit_insns_count:      1780584955
ratio_in_yjit:              86.1%
avg_len_in_yjit:             27.5
Top-20 most frequent exit ops (100.0% of exits):
    opt_send_without_block:   10451094 (45.6%)
                      send:    6600826 (28.8%)
               invokesuper:    2763908 (12.1%)
               invokeblock:    1603119 (7.0%)
               getconstant:     432849 (1.9%)
                  opt_aref:     350087 (1.5%)
                     throw:     259112 (1.1%)
               expandarray:     249254 (1.1%)
             setlocal_WC_0:     149541 (0.7%)
      opt_getconstant_path:      32602 (0.1%)
               opt_empty_p:        744 (0.0%)
       getinstancevariable:        415 (0.0%)
        getblockparamproxy:        244 (0.0%)
                     leave:        210 (0.0%)
                checkmatch:        193 (0.0%)
               objtostring:        138 (0.0%)
                 opt_nil_p:         30 (0.0%)
          opt_newarray_max:         29 (0.0%)
                      once:         27 (0.0%)
             setblockparam:         15 (0.0%)
Total time spent benchmarking: 37s

end_time: 2022-09-16 13:28:25 JST (+0900)
/Users/k0kubun/.rbenv/versions/yjit-arm64/bin/ruby --yjit-stats: ruby 3.2.0dev (2022-09-15T20:19:50Z splat-args-fix e2a641724a) +YJIT [arm64-darwin21]

----------  --------------------------------------------------------------------  ----------
bench       /Users/k0kubun/.rbenv/versions/yjit-arm64/bin/ruby --yjit-stats (ms)  stddev (%)
railsbench  1236.8                                                                1.3
----------  --------------------------------------------------------------------  ----------
```

</details>

`args_splat_non_iseq      50435 ( 0.3%)` newly appeared at method call exit reasons, but its impact on overall opt_send_without_block/send exit counts seems minimal. Let me merge this :pray: